pyproject.toml
[tool.poetry]
name = "mosaic"
version = "1.5.0"
description = "A tool to convert images into a mosaic of fuzzy, diffuse gray shapes with video generation support."
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "mosaic", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
Pillow = "^10.0.0"
opencv-python = "^4.13.0.92"
numpy = "^2.4.2"
tqdm = "^4.66.0"
ai-api-unified = {extras = ["google-gemini"], version = "^1.3.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[tool.poetry.scripts]
mosaic = "mosaic.mosaic_generator:main"
mosaic-test = "pytest:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
